ТЗ 15

+++ src/main/java/ru/practicum/shareit/user/service/UserService.java
@@ -1,18 +1,17 @@
 package ru.practicum.shareit.user.service;
 
 import ru.practicum.shareit.user.User;
-
 import java.util.List;
 
 public interface UserService {
 
     User create(User user);
 
-    User update(int userId, User user);
+    User update(Long userId, User user);
 
-    User getById(int userId);
+    User getById(Long userId);
 
     List<User> getAll();
 
-    void delete(int userId);
+    void delete(Long userId);
 }
